Move CSSStyledText to braille-css-utils
