+++ braille/dotify-utils/src/test/java/DotifyCoreTest.java
@@ -5,9 +5,9 @@ import javax.inject.Inject;
 
 import org.daisy.braille.css.SimpleInlineStyle;
 
-import org.daisy.pipeline.braille.common.CSSStyledText;
 import static org.daisy.pipeline.braille.common.Query.util.query;
 import org.daisy.pipeline.braille.dotify.DotifyTranslator;
+import org.daisy.pipeline.braille.css.CSSStyledText;
 import org.daisy.pipeline.junit.AbstractTest;
 
 import org.junit.Test;
