+++ braille/liblouis-utils/src/test/java/LiblouisTablesTest.java
@@ -7,8 +7,8 @@ import javax.inject.Inject;
 
 import com.google.common.base.Optional;
 
-import org.daisy.pipeline.braille.common.CSSStyledText;
 import static org.daisy.pipeline.braille.common.Query.util.query;
+import org.daisy.pipeline.braille.css.CSSStyledText;
 import org.daisy.pipeline.braille.liblouis.LiblouisTranslator;
 
 import org.daisy.pipeline.junit.AbstractTest;
